Fix bug with restarts by saving the optimizer state
The problem was twofold:
1/ the optimizer weights were not preserved across restarts so the optimizer started in a wrong place
2/ the optimizer iterations were not saved so the learning rate, the decay of which is computed from the iterations, was incorrect.

This all comes from the TF bug (apparently) whereby the optimizer is *not* saved (despite what the documentation says) when saving in the native TF format, while it is when saving in the HDF5 format.  This commit makes us use the HDF5 format for the checkpoints.